Improved BuildModeModule (update to the newest Darkness standards)
