+++ modules/Buildmode/src/main/resources/module.properties
@@ -1,1 +1,1 @@
-main=de.astride.servercore.modules.buildmode.BuildmodeModule
+main=de.astride.servercore.modules.buildmode.BuildModeModule
